migrate to ApplicationUser
migrates the application, as well as the database, to be based on
ApplicationUser rather than the default IdentityUser; this change mostly
is about changing the user's primary key to a Guid type rather than
the default string. for this, ApplicationUser inherits from the generic
version of IdentityUser<>, which allows the choice of Guid.